requirements.txt
discord.py
python-dotenv
aiohttp
aiosqlite
watchdog
prometheus_client
psutil
pillow
numpy
wikipedia
aiomysql
edge_tts
cryptography
PyNaCl
matplotlib
scikit-learn
prophet
pandas